requirements.txt
Django==3.2.6
django-cors-headers==3.5.0
django-storages==1.12.2
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
